progress on ensuring Java's long is represented as int64 (for #83)

+++ database_put_result.go
@@ -2,7 +2,7 @@ package ravendb
 
 // DatabasePutResult describes server response for e.g. CreateDatabaseCommand
 type DatabasePutResult struct {
-	RaftCommandIndex int      `json:"RaftCommandIndex"`
+	RaftCommandIndex int64    `json:"RaftCommandIndex"`
 	Name             string   `json:"Name"`
 	DatabaseTopology Topology `json:"Topology"`
 	NodesAddedTo     []string `json:"NodesAddedTo"`
