+++ database_record_with_etag.go
@@ -3,5 +3,5 @@ package ravendb
 // DatabaseRecordWithEtag represents database record with etag
 type DatabaseRecordWithEtag struct {
 	DatabaseRecord
-	Etag int `json:"Etag"`
+	Etag int64 `json:"Etag"`
 }
