+++ leader_stamp.go
@@ -2,7 +2,7 @@ package ravendb
 
 // LeaderStamp describes leader stamp
 type LeaderStamp struct {
-	Index        int `json:"Index"`
-	Term         int `json:"Term"`
-	LeadersTicks int `json:"LeadersTicks"`
+	Index        int64 `json:"Index"`
+	Term         int64 `json:"Term"`
+	LeadersTicks int64 `json:"LeadersTicks"`
 }
